Fixed Loading Game Score Bug

See merge request compsci308_2024spring/oogasalad_team01!142

+++ src/main/java/oogasalad/model/api/PlayerRecord.java
@@ -9,6 +9,6 @@ package oogasalad.model.api;
  *                         strikeable
  * @author Noah Loewy
  */
-public record PlayerRecord(int playerId, double score, int activeStrikeable) {
+public record PlayerRecord(int playerId, int score, int activeStrikeable) {
 }
